+++ CifraShop.Contracts/Responses/Products/ProductResponse.cs
@@ -7,8 +7,8 @@ namespace CifraShop.Contracts.Responses.Products
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public short Price { get; set; }
-        public short Quantity { get; set; }
+        public int Price { get; set; }
+        public int Quantity { get; set; }
         public StatusProduct Status { get; set; }
         public string? ImageUrl { get; set; }
     }
