+++ CifraShop.Domain/Entities/User.cs
@@ -7,7 +7,7 @@ namespace CifraShop.Domain.Entities
         public int Id { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public short? Balance { get; set; }
+        public int? Balance { get; set; }
         public UserRole Role { get; set; }
         public string? Branch { get; set; }
         public List<Order> Orders { get; set; } = new List<Order>();
